Refactor web service
