+++ requirements.txt
@@ -1,8 +1,7 @@
-Flask==0.12
+Flask==0.12.3
 tensorflow==0.12.0
 Pillow==3.4.2
-matplotlib==1.5.3
 numpy==1.11.3
-uwsgi==2.0.14
+uwsgi==2.0.15
 celery==4.0.1
-pytest
+coverage==4.5.2
