+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup
+
+setup(
+  name = 'blueberry',
+  packages = ['blueberry'],
+  include_package_data = True,
+  install_requires = ['flask']
+)
