+++ uwsgi.ini
@@ -1,8 +1,8 @@
 [uwsgi]
 http-socket = :80
 wsgi-file = wsgi.py
-module = main
+module = blueberry.main
 callable = app
 master = true
-processes = 2
+processes = 1
 die-on-term = true
