+++ wsgi.py
@@ -1,4 +1,4 @@
-from main import app
+from blueberry.main import app
 
 if __name__ == "__main__":
   app.run(host = "0.0.0.0", port = 80)
